pyproject.toml
[tool.poetry]
name = "remote-ipython"
version = "0.1.0a4"
description = ""
authors = ["likianta <likianta@foxmail.com>"]
# readme = "README.md"
packages = [{ include = "remote_ipython" }]

[tool.poetry.dependencies]
python = "^3.8"
ipykernel = "^6.27.1"
ipython = "^8.12.0"
# jupyter = "^1.0.0"
jupyter-console = "^6.6.3"
lk-logger = { version = "^6.0.0a1", source = "likianta-host" }
lk-utils = { version = "^2.10.8", source = "likianta-host" }
prompt-toolkit = "^3.0.41"
pyzmq = "^26.0.3"
# remote-kernel = "^1.1"
# rich = "*"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "likianta-host"
url = "http://likianta.pro:2006/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
